Fix auto indent when pasting lines that begin with a comment
The special handling for comments was originally added in 886f472
to fix #29753, but it had the side effect of preventing a desired auto
indent in the majority of cases when pasting lines beginning with a
comment (basically, all cases except multiline docblock style comments
at the leftmost indentation level). That case has since been solved in
a different way by e77b196, so removing this special handling gives us
the best of both worlds.